flowork-gateway/requirements.txt
#######################################################################
# WEBSITE https://flowork.cloud
# File NAME : C:\FLOWORK\flowork-gateway\requirements.txt
#######################################################################
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
python-json-logger==2.0.7
SQLAlchemy==2.0.23
web3==6.11.1
gunicorn==21.2.0
requests==2.31.0
Flask-Cors
python-socketio==5.12.1
msgpack==1.0.7
gevent
